+++ CHANGELOG.md
@@ -726,6 +726,7 @@
 - Delivery partner users have a home page that lets them view and search their
   own activities
 - Associate Report with HistoricalEvent when uploading a bulk CSV
+- Add a report summary tab
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-59...HEAD
 [release-59]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-58...release-59